+++ src/ts/VAST/Utilities/VastParser.ts
@@ -307,6 +307,7 @@ export class VastParser {
     private parseCreativeCompanionAdElement(companionAdElement: Element): VastCreativeStaticResourceCompanionAd | null {
         const staticResourceElement = <Element>this.childByName(companionAdElement, 'StaticResource');
         const companionClickThroughElement = this.childByName(companionAdElement, 'CompanionClickThrough');
+        const companionClickTrackingElements = <Element[]>this.childsByName(companionAdElement, 'CompanionClickTracking');
 
         if (companionAdElement && staticResourceElement) {
             const id = companionAdElement.getAttribute('id');
@@ -315,10 +316,16 @@ export class VastParser {
             const creativeType = staticResourceElement.getAttribute('creativeType');
             const staticResourceURL = this.parseNodeText(staticResourceElement);
             const companionClickThroughURLTemplate = this.parseNodeText(companionClickThroughElement!);
-
+            const companionClickTrackingURLTemplates: string[] = [];
+            for (const trackingElement of companionClickTrackingElements) {
+                const companionClickTrackingTemplate = this.parseNodeText(trackingElement);
+                if (companionClickTrackingTemplate != null) {
+                    companionClickTrackingURLTemplates.push(companionClickTrackingTemplate);
+                }
+            }
             const trackingEvents = this.getTrackingEventsFromElement(companionAdElement);
 
-            return new VastCreativeStaticResourceCompanionAd(id, height, width, creativeType, staticResourceURL, companionClickThroughURLTemplate, trackingEvents);
+            return new VastCreativeStaticResourceCompanionAd(id, height, width, creativeType, staticResourceURL, companionClickThroughURLTemplate, companionClickTrackingURLTemplates, trackingEvents);
         } else {
             return null;
         }